Changing About us list order style
